Fix: Discrete actions with size > 2 showing only binary values in Godot
Discrete actions with size > 2 would display only 0 or 1 even when using the multi discrete spaces.

This fix addresses that by removing the binary conversion (np.greater) in case of using multi discrete spaces. 

A user-friendly error message is also added when trying to mix a discrete action larger than 2 with continuous actions. 

I have quickly tested this on modified "RingPong" example code and it works in cases of: 
1) 1 Discrete action with size > 2. 
2) 2 Discrete actions with size > 2.
2) Discrete action with size up to 2 + continuous action.

However, as I'm not familiar with the code I don't know if this is the right solution, or if it might break something else.